impl avx2, neon private headers
