+++ CMakeLists.txt
@@ -137,8 +137,10 @@ endif()
 set(libspeck_SRCS ${libspeck_SRCS} src/speck/speck_common.c)
 if(WITH_NEON)
     set(libspeck_SRCS ${libspeck_SRCS} src/speck/speck_neon.c)
+    set(libspeck_PRIVATE_HEADER ${libspeck_PRIVATE_HEADER} src/speck/speck_neon_private.h)
 elseif(WITH_AVX2)
     set(libspeck_SRCS ${libspeck_SRCS} src/speck/speck_avx2.c)
+    set(libspeck_PRIVATE_HEADER ${libspeck_PRIVATE_HEADER} src/speck/speck_avx2_private.h)
 else()
     set(libspeck_SRCS ${libspeck_SRCS} src/speck/speck.c)
 endif()
@@ -151,7 +153,7 @@ endif(WIN32)
 
 # header
 set(libspeck_PUBLIC_HEADER include/speck/speck.h)
-set(libspeck_PRIVATE_HEADER src/speck/speck_private.h src/speck/speck_ctr_private.h)
+set(libspeck_PRIVATE_HEADER ${libspeck_PRIVATE_HEADER} src/speck/speck_private.h src/speck/speck_ctr_private.h)
 
 include_directories(${CMAKE_SOURCE_DIR}/include)
 set(libspeck_SRCS ${libspeck_SRCS} ${libspeck_PUBLIC_HEADER} ${libspeck_PRIVATE_HEADER})
